+++ lib/kb_gtdbtk/core/api_translation.py
@@ -54,9 +54,8 @@ def get_gtdbtk_params(input_params: Dict[str, object]) -> GTDBTKParams:
     if type(wsid) != int or _cast(int, wsid) < 1:
         raise ValueError('workspace_id is required and must be an integer > 0')
 
-    overwrite_tax = input_params.get('overwrite_tax', 1)
-    if type(overwrite_tax) != int \
-       or (_cast(int, overwrite_tax) != 0 and _cast(int, overwrite_tax) != 1):
+    overwrite_tax = _cast(int, input_params.get('overwrite_tax', 1))
+    if type(overwrite_tax) != int or (overwrite_tax != 0 and overwrite_tax != 1):
         raise ValueError('overwrite_tax is required and must be an integer [0,1]')